pyproject.toml
[project]
name = "seat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ddddocr>=1.6.1",
    "playwright>=1.58.0",
    "pycryptodome>=3.23.0",
]
